+++ src/Store.php
@@ -123,7 +123,7 @@ final class Store {
 		}
 
 		if ( ! get_userdata( $user_id ) ) {
-			return new \WP_Error( 'user_tags_unknown_user', __( 'That user does not exist.', 'user-tags' ) );
+			return new \WP_Error( 'user_tags_unknown_user', __( 'That user does not exist.', 'user-tags-lib' ) );
 		}
 
 		$catalogue = Catalogue::all();